\name{bayesmeta-package}
\alias{BayesMeta}
\docType{package}
\title{
\packageTitle{bayesmeta}
}
\description{
Estimates consensus age-specific penetrance of cancer for carriers of a gene mutation integrating studies reporting age-specific penetrance, relative risk, odds
ratio and  standardized incidence ratio through a  approach based on a Bayesian hierarchical random-effects model.
}
\details{

This package contains a function called \code{'bayesmeta'} for meta analysis to estimate consensus age-specific penetrance through Bayesian hierarchical random-effects model
}